+++ CHANGELOG.md
@@ -43,3 +43,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Implement Token API, [PR-15](https://github.com/reductstore/reduct-go/pull/15)
 * Implement Replication API, [PR-16](https://github.com/reductstore/reduct-go/pull/16)
 * Improvement for Bucket API, [PR-18](https://github.com/reductstore/reduct-go/pull/18)
+* Improvement for Bucket API, [PR-30](https://github.com/reductstore/reduct-go/pull/30)